DeltaTime and Updates first steps
Actual Version: v1.0.0.2
Minor Changes: Improved the StaticElementsLayer drawing function: Now
there is no need for loading the sprite image at each step of the loop.
Major Changes: Splitted the drawing function into two: One for updating
the changes of the layer elements and another for the drawing.
Started the integration of the deltaTime concept for the gameLoop speed
consistency.